pyproject.toml
[project]
name = "smart-formatting-llm-benchmark"
version = "0.1.0"
description = "OpenRouter benchmark runner + evaluator for ASR transcript formatting"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pandas>=2.2.0",
    "anthropic>=0.39",
    "openai>=1.50",
    "google-genai>=0.5",
    "scipy>=1.11.0",
    "rapidfuzz>=3.9.0",
]

[project.scripts]
runner = "runner.main:app"
evaluator = "evaluator.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["runner", "evaluator"]
